更改蓄电池电压数据类型

+++ Yamaha/src/CustomCarMsg.cpp
@@ -602,7 +602,7 @@ void CustomCarMsg::updateMoterControlTemp(int value) {
 void CustomCarMsg::updateMoterTemp(int value) {
     MEMBER_PROPERTY_VALUE_CHANGED(moterTemp, value);
 }
-void CustomCarMsg::updateBattery(float value) {
+void CustomCarMsg::updateBattery(int value) {
     qDebug() << "1111111" << value;
     MEMBER_PROPERTY_VALUE_CHANGED(battery, value);
 }
